Add managed packages & remote generation documentation (#220)
Behold - "remote packages/plugins" documentation... The goal of this is to try
and provide a brand for our customers to distinguish between our
different remote generation offerings and provide clarity in
communication moving forward. A major issue we've found internally is,
as the conversation of "remote generation" has widened, the team has
struggled to grasp exactly what the feature is and how it differs from
"Templates".

This poses a significant problem, if we can't effectively communicate
amongst ourselves, how can we expect the unindoctrinated to be able to
see how "remote generation" is going to help them in their personal
projects, startups or companies.

From this problem, we propose a solution... to divide our features on
customer value - a feature which results in code and those another which
results in ready to use packages.
1. Generate - used to supply proto files and retrieve generated code.
Broken down into 2 parts:
    - Local Generation - utilize your locally installed protoc plugins 
- Remote Generation - The Buf team manages remote protobuf plugins that
can be referenced in `buf.gen.yaml` files. Code generation takes place
remotely on the BSR and generated source code is written out to disk.
2. Remote Packages - The BSR exposes generated source code through
managed software repositories you fetch like any other library with
tools you already know: `go get` or `npm install`.

Co-authored-by: bufdev <4228796+bufdev@users.noreply.github.com>
Co-authored-by: Joshua Carpeggiani <32605850+joshcarp@users.noreply.github.com>
Co-authored-by: Philip K. Warren <pkwarren@users.noreply.github.com>
Co-authored-by: Saquib Mian <saquib.mian@gmail.com>
Co-authored-by: bufdev <bufdev-github@buf.build>
Co-authored-by: Ollie Hsieh <ohsieh@buf.build>
Co-authored-by: Mike Fridman <mfridman@buf.build>
Co-authored-by: Philip K. Warren <pwarren@buf.build>